+++ DATN.Application/Dtos/TestSetDtos/TestSetDetailsDto.cs
@@ -22,6 +22,6 @@ namespace DATN.Application.Dtos.TestSetDtos
         public int CountUserDo { get; set; }
         public string RankQuestionName { get; set; }
 
-        public int AvgRating { get; set; }
+        public double AvgRating { get; set; }
     }
 }
